move verifyHasValidScopes to separate file
