+++ src/middleware/graphql/graphqlServer.js
@@ -50,7 +50,7 @@ const graphql = async () => {
                 const requestInfo = {
                     user: (req.authInfo && req.authInfo.context && req.authInfo.context.username) ||
                         (req.authInfo && req.authInfo.context && req.authInfo.context.subject) ||
-                        req.user,
+                        ((!req.user || typeof req.user === 'string') ? req.user : req.user.id),
                     patients: req.authInfo && req.authInfo.context && req.authInfo.context.fhirPatientIds,
                     scope: req.authInfo && req.authInfo.scope,
                     remoteIpAddress: req.headers['x-forwarded-for'] || req.connection.remoteAddress,
