+++ src/operations/security/scopes.js
@@ -1,8 +1,7 @@
 const env = require('var');
-const scopeChecker = require('@asymmetrik/sof-scope-checker');
+const assert = require('node:assert/strict');
 const {ForbiddenError} = require('../../utils/httpErrors');
-const {logOperation} = require('../common/logging');
-const {authorizationFailedCounter} = require('../../utils/prometheus.utils');
+
 /**
  * converts a space separated list of scopes into an array of scopes
  * @param {string} scope
@@ -16,76 +15,15 @@ const parseScopes = (scope) => {
 };
 
 /**
- * @typedef VerifyScopesParameters
- * @type {object}
- * @property {import('../../utils/requestInfo').RequestInfo} requestInfo
- * @property {Object} args
- * @property {string} resourceType
- * @property {number|null} startTime
- * @property {string} action
- * @property {string} accessRequested
- */
-
-/**
- * Throws an error if no scope is valid for this request
- * @param {VerifyScopesParameters} options
- */
-const verifyHasValidScopes = (options) => {
-    const {requestInfo, args, resourceType, startTime, action, accessRequested} = options;
-    const {user, scope} = requestInfo;
-
-    if (env.AUTH_ENABLED === '1') {
-        // http://www.hl7.org/fhir/smart-app-launch/scopes-and-launch-context/index.html
-        if (scope) {
-            /**
-             * @type {string[]}
-             */
-            let scopes = parseScopes(scope);
-            let {error, success} = scopeChecker(resourceType, accessRequested, scopes);
-
-            if (success) {
-                return;
-            }
-            let errorMessage = 'user ' + user + ' with scopes [' + scopes + '] failed access check to [' + resourceType + '.' + accessRequested + ']';
-            const forbiddenError = new ForbiddenError(error.message + ': ' + errorMessage);
-            authorizationFailedCounter.inc({action: action, resourceType: resourceType});
-            logOperation({
-                requestInfo,
-                args,
-                resourceType,
-                startTime: startTime,
-                message: 'AuthorizationFailed',
-                action: action,
-                error: forbiddenError
-            });
-            throw forbiddenError;
-        } else {
-            let errorMessage = 'user ' + user + ' with no scopes failed access check to [' + resourceType + '.' + accessRequested + ']';
-            const forbiddenError1 = new ForbiddenError(errorMessage);
-            authorizationFailedCounter.inc({action: action, resourceType: resourceType});
-            logOperation({
-                requestInfo,
-                args,
-                resourceType,
-                startTime: startTime,
-                message: 'AuthorizationFailed',
-                action: action,
-                error: forbiddenError1
-            });
-            throw forbiddenError1;
-        }
-    }
-};
-
-/**
  * Returns all the access codes present in scopes
  * @param {string} action
  * @param {string} user
- * @param {?string} scope
- * @return {string[]}
+ * @param {string|null} scope
+ * @return {string[]} security tags allowed by scopes
  */
 const getAccessCodesFromScopes = (action, user, scope) => {
     if (env.AUTH_ENABLED === '1') {
+        assert(typeof user === 'string', `user is of type: ${typeof user} but should be string.`);
         // http://www.hl7.org/fhir/smart-app-launch/scopes-and-launch-context/index.html
         /**
          * @type {string[]}
@@ -105,16 +43,9 @@ const getAccessCodesFromScopes = (action, user, scope) => {
                  * @type {string}
                  */
                 const inner_scope = scope1.replace('access/', '');
-                /**
-                 * @type {[string]}
-                 */
-                const innerScopes = inner_scope.split('.');
-                if (innerScopes && innerScopes.length > 0) {
-                    /**
-                     * @type {string}
-                     */
-                    const access_code = innerScopes[0];
-                    access_codes.push(access_code);
+                const [securityTag, accessType] = inner_scope.split('.');
+                if (accessType === '*' || accessType === action) {
+                    access_codes.push(securityTag);
                 }
             }
         }
@@ -209,7 +140,6 @@ const doesResourceHaveAccessTags = (resource) => {
 module.exports = {
     isAccessToResourceAllowedBySecurityTags,
     doesResourceHaveAccessTags,
-    verifyHasValidScopes,
     getAccessCodesFromScopes,
     doesResourceHaveAnyAccessCodeFromThisList,
     parseScopes
